net-im/nheko(-reborn): re-add
This adds Nheko back to the repository which was removed because the old
maintainer didn't have that much time to maintain it anymore. Since then
development resumed in the Nheko-Reborn repository and development is
somewhat active.

Currently this doesn't build with the upstream nlohmann_json and it has
some issues with const variables used in a non-const context, so there
are some manual patches needed.